enhanced new animation test
